Revert "feat的功能是增加了选择JSON文件的功能，解析JSON数据并更新Excel文件中的车位状态。具体改动如下：- 在ControlPanel中添加了选择文件按钮"
This reverts commit a88f848c1d338fcc2692df31b0f43d20761347d1.

+++ src/main/java/com/yuda/parking/HousewindowList.java
@@ -25,7 +25,7 @@ public class HousewindowList {
     private double otherarea;
     private long permitid;
     private String permitno;
-    private int presaleid;
+    private long presaleid;
     private long price;
     private long prjid;
     private String projectname;
